+++ templates/portal/base.html
@@ -7,16 +7,16 @@
 
     <link href="{% static 'fontawesomefree/css/all.min.css' %}" rel="stylesheet" type="text/css">
     <script src="{% static 'fontawesomefree/js/all.min.js' %}"></script>
-    <script src="{% static 'portal/js/jquery/jquery-3.6.0.min.js' %}"></script>
+    <script src="{% static 'portal/../../static/users/js/jquery/jquery-3.6.0.min.js' %}"></script>
 
 
-    <link href="{% static 'portal/css/animations.css' %}" rel="stylesheet" type="text/css">
-    <link href="{% static 'portal/css/base.css' %}" rel="stylesheet" type="text/css">
-    <link href="{% static 'portal/css/navbar.css' %}" rel="stylesheet" type="text/css">
-    <link href="{% static 'portal/css/form.css' %}" rel="stylesheet" type="text/css">
-    <link href="{% static 'portal/css/modal.css' %}" rel="stylesheet" type="text/css">
-    <link href="{% static 'portal/css/footer.css' %}" rel="stylesheet" type="text/css">
-    <link href="{% static 'portal/css/responsive.css' %}" rel="stylesheet" type="text/css">
+    <link href="{% static 'portal/../../static/users/css/animations.css' %}" rel="stylesheet" type="text/css">
+    <link href="{% static 'portal/../../static/users/css/base.css' %}" rel="stylesheet" type="text/css">
+    <link href="{% static 'portal/../../static/users/css/navbar.css' %}" rel="stylesheet" type="text/css">
+    <link href="{% static 'portal/../../static/users/css/form.css' %}" rel="stylesheet" type="text/css">
+    <link href="{% static 'portal/../../static/users/css/modal.css' %}" rel="stylesheet" type="text/css">
+    <link href="{% static 'portal/../../static/users/css/footer.css' %}" rel="stylesheet" type="text/css">
+    <link href="{% static 'portal/../../static/users/css/responsive.css' %}" rel="stylesheet" type="text/css">
     <link rel="icon" href="{% static 'img/favicon.ico' %}">
 
 
@@ -25,7 +25,7 @@
 <div class="ov-content">
     <nav class="ov-nav">
         <div class="ov-nav__left">
-            <img src="{% static 'portal/images/logo_green.png' %}" alt="Planet Soft Logo" class="ov-nav__logo">
+            <img src="{% static 'portal/../../static/users/images/logo_green.png' %}" alt="Planet Soft Logo" class="ov-nav__logo">
             <h1 class="ov-nav__title">PLANET SOFT</h1>
         </div>
         <div class="ov-nav__right">
@@ -90,9 +90,9 @@
     </div>
 </footer>
 
-<script src="{% static 'portal/js/base.js' %}"></script>
+<script src="{% static 'portal/../../static/users/js/base.js' %}"></script>
 
-<script src="{% static 'portal/js/modal-utils.js' %}"></script>
+<script src="{% static 'portal/../../static/users/js/modal-utils.js' %}"></script>
 
 
 </body>
