--- templates/portal/templates/employee/employee-detail.html
@@ -1,1 +1,0 @@
-<h1>{{ employee.first_name }} {{ employee.last_name }}</h1>
