+++ templates/users/employee/employee-edit.html
@@ -1,7 +1,7 @@
 <form id="edit-emp-form" enctype="multipart/form-data" class="ov-form">
     <h3 class="ov-form__title">Uređivanje postojećeg korisnika</h3>
 
-{% if form %}
+{% if not form_preview %}
     {% csrf_token %}
     {% for error in form.non_field_errors %}
             <div>
